+++ examples/preprocessor/src/includes/include_b.js
@@ -1,6 +1,6 @@
 {
 	const $content = document.getElementById('content');
 	const $p = document.createElement('p');
-	$p.innerHTML = 'Output from <span class="highlight">included_a.js</span>!';
+	$p.innerHTML = 'This is output from included script at <span class="highlight">include\\include_b.js</span>!';
 	$content.appendChild($p);
 }